refactor(codegen): extract call lowering helpers
