+++ crates/codegen/src/lower/expr.rs
@@ -8,10 +8,9 @@ use crate::{
     mir::{FunctionBuilder, MirType, ValueId},
     utils::{ConstantFolder, FoldResult},
 };
-use alloy_primitives::{U256, keccak256};
+use alloy_primitives::U256;
 use solar_ast::{LitKind, StrKind};
-use solar_data_structures::map::FxHashSet;
-use solar_interface::{Ident, Span, Symbol, kw, sym};
+use solar_interface::{Ident, Span, sym};
 use solar_sema::{
     builtins::Builtin,
     hir::{self, CallArgs, ElementaryType, ExprKind},
@@ -764,7 +763,7 @@ impl<'gcx> Lowerer<'gcx> {
         builder.phi(incoming)
     }
 
-    fn emit_revert_error_string_from_expr(
+    pub(super) fn emit_revert_error_string_from_expr(
         &mut self,
         builder: &mut FunctionBuilder<'_>,
         expr: &hir::Expr<'_>,
@@ -1208,229 +1207,7 @@ impl<'gcx> Lowerer<'gcx> {
         }
     }
 
-    /// Lowers a function call.
-    fn lower_call(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        callee: &hir::Expr<'_>,
-        args: &CallArgs<'_>,
-        call_opts: Option<&[hir::NamedArg<'_>]>,
-    ) -> ValueId {
-        if let ExprKind::Ident(res_slice) = &callee.kind
-            && let Some(builtin) = self.select_builtin_overload(res_slice, args)
-        {
-            return self.lower_builtin_call(builder, builtin, args);
-        }
-
-        if let Some(error_id) = self.custom_error_id_from_callee(callee) {
-            self.emit_custom_error_revert(builder, error_id, args);
-            return builder.imm_u64(0);
-        }
-
-        // `T.wrap(x)` / `T.unwrap(v)` for a user-defined value type are identity
-        // operations at the EVM level: a UDVT value is represented exactly as its
-        // underlying type, so no wrapper is added or removed.
-        if let ExprKind::Member(base, member) = &callee.kind
-            && matches!(member.name.as_str(), "wrap" | "unwrap")
-            && let ExprKind::Ident(res_slice) = &base.kind
-            && res_slice.iter().any(|r| matches!(r, hir::Res::Item(hir::ItemId::Udvt(_))))
-            && let Some(arg) = args.exprs().next()
-        {
-            return self.lower_expr(builder, arg);
-        }
-
-        if let ExprKind::Member(base, member) = &callee.kind {
-            return self.lower_member_call_with_opts(builder, base, *member, args, call_opts);
-        }
-
-        // Handle `new Contract(args)` - contract creation
-        if let ExprKind::New(ty) = &callee.kind {
-            if self.is_memory_array_new_type(ty) {
-                return self.lower_new_array(builder, ty, args);
-            }
-            return self.lower_new_contract(builder, ty, args, call_opts);
-        }
-
-        // Handle internal function calls: func(args) where func is a function in the same contract
-        if let ExprKind::Ident(res_slice) = &callee.kind {
-            let arg_count = args.exprs().count();
-
-            // Find the best matching overload based on argument count
-            // First, collect all function resolutions
-            let func_resolutions: Vec<_> = res_slice
-                .iter()
-                .filter_map(|r| {
-                    if let hir::Res::Item(hir::ItemId::Function(fid)) = r {
-                        Some(*fid)
-                    } else {
-                        None
-                    }
-                })
-                .collect();
-
-            // Select the overload that matches the argument count
-            let selected_func = func_resolutions
-                .iter()
-                .find(|&&fid| {
-                    let f = self.gcx.hir.function(fid);
-                    f.parameters.len() == arg_count
-                })
-                .or_else(|| func_resolutions.first());
-
-            if let Some(&func_id) = selected_func {
-                return self.lower_internal_call(builder, func_id, args);
-            }
-        }
-
-        // Handle type conversion calls: Type(expr)
-        // e.g., ICallee(addr), uint256(x), address(y), Status(n)
-        // The callee is an Ident resolving to a contract/interface/enum type
-        if let ExprKind::Ident(res_slice) = &callee.kind {
-            // Check if this resolves to a contract or interface type
-            if let Some(hir::Res::Item(hir::ItemId::Contract(_))) = res_slice.first() {
-                // Type conversion: just return the first argument unchanged
-                // (The actual conversion is a no-op at the EVM level for addresses/contracts)
-                if let Some(first_arg) = args.exprs().next() {
-                    return self.lower_expr(builder, first_arg);
-                }
-            }
-            // Check if this resolves to an enum type (e.g., Status(n))
-            if let Some(hir::Res::Item(hir::ItemId::Enum(_))) = res_slice.first() {
-                // Enum conversion: just return the first argument unchanged
-                // (Enums are represented as uint8 at the EVM level)
-                if let Some(first_arg) = args.exprs().next() {
-                    return self.lower_expr(builder, first_arg);
-                }
-            }
-            // Check if this resolves to a struct type (struct constructor call)
-            // e.g., Point(10, 20) creates a memory struct
-            if let Some(hir::Res::Item(hir::ItemId::Struct(struct_id))) = res_slice.first() {
-                return self.lower_struct_constructor(builder, *struct_id, args);
-            }
-        }
-
-        // Handle Type(expr) where callee is an explicit Type expression
-        // e.g., uint256(x), address(y), bytes32(z)
-        if let ExprKind::Type(ty) = &callee.kind
-            && let Some(first_arg) = args.exprs().next()
-        {
-            let value = self.lower_expr(builder, first_arg);
-            return self.lower_type_conversion(builder, ty, first_arg, value);
-        }
-
-        builder.imm_u64(0)
-    }
-
-    fn select_builtin_overload(
-        &self,
-        res_slice: &[hir::Res],
-        args: &CallArgs<'_>,
-    ) -> Option<Builtin> {
-        let arg_count = args.exprs().count();
-        res_slice
-            .iter()
-            .filter_map(|res| match res {
-                hir::Res::Builtin(builtin) => Some(*builtin),
-                _ => None,
-            })
-            .find(|builtin| match builtin {
-                Builtin::Revert => arg_count == 0,
-                Builtin::RevertMsg => arg_count == 1,
-                _ => true,
-            })
-    }
-
-    fn custom_error_id_from_callee(&self, callee: &hir::Expr<'_>) -> Option<hir::ErrorId> {
-        if let Some(ty) = self.get_expr_type(callee)
-            && let TyKind::Error(_, error_id) = ty.kind
-        {
-            return Some(error_id);
-        }
-
-        let ExprKind::Ident(res_slice) = &callee.kind else { return None };
-        res_slice.iter().find_map(|res| {
-            if let hir::Res::Item(hir::ItemId::Error(error_id)) = res {
-                Some(*error_id)
-            } else {
-                None
-            }
-        })
-    }
-
-    fn emit_revert_payload_from_expr(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        expr: &hir::Expr<'_>,
-    ) -> bool {
-        if self.emit_custom_error_revert_from_expr(builder, expr) {
-            return true;
-        }
-        self.emit_revert_error_string_from_expr(builder, expr)
-    }
-
-    fn emit_custom_error_revert_from_expr(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        expr: &hir::Expr<'_>,
-    ) -> bool {
-        let ExprKind::Call(callee, args, _) = &expr.kind else { return false };
-        let Some(error_id) = self.custom_error_id_from_callee(callee) else {
-            return false;
-        };
-        self.emit_custom_error_revert(builder, error_id, args);
-        true
-    }
-
-    fn emit_custom_error_revert(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        error_id: hir::ErrorId,
-        args: &CallArgs<'_>,
-    ) {
-        let param_tys = self.gcx.item_parameter_types(hir::ItemId::Error(error_id));
-        let arg_exprs = self.ordered_custom_error_args(error_id, args);
-        let mut items = Vec::with_capacity(param_tys.len());
-        for (&ty, arg) in param_tys.iter().zip(arg_exprs) {
-            let value = self.lower_return_value_for_ty(builder, arg, ty);
-            items.push((value, ty));
-        }
-
-        let selector = self.custom_error_selector(error_id);
-        self.emit_abi_error_revert(builder, selector, &items);
-    }
-
-    fn ordered_custom_error_args<'a>(
-        &self,
-        error_id: hir::ErrorId,
-        args: &'a CallArgs<'a>,
-    ) -> Vec<&'a hir::Expr<'a>> {
-        match args.kind {
-            hir::CallArgsKind::Unnamed(exprs) => exprs.iter().collect(),
-            hir::CallArgsKind::Named(named_args) => {
-                let error = self.gcx.hir.error(error_id);
-                let mut ordered = Vec::with_capacity(error.parameters.len());
-                for &param_id in error.parameters {
-                    let Some(param_name) =
-                        self.gcx.hir.variable(param_id).name.map(|name| name.name)
-                    else {
-                        continue;
-                    };
-                    if let Some(arg) = named_args.iter().find(|arg| arg.name.name == param_name) {
-                        ordered.push(&arg.value);
-                    }
-                }
-                ordered
-            }
-        }
-    }
-
-    fn custom_error_selector(&self, error_id: hir::ErrorId) -> [u8; 4] {
-        let signature = self.gcx.item_signature(hir::ItemId::Error(error_id));
-        let hash = keccak256(signature.as_bytes());
-        [hash[0], hash[1], hash[2], hash[3]]
-    }
-
-    fn lower_type_conversion(
+    pub(super) fn lower_type_conversion(
         &mut self,
         builder: &mut FunctionBuilder<'_>,
         ty: &hir::Type<'_>,
@@ -1563,1059 +1340,6 @@ impl<'gcx> Lowerer<'gcx> {
         builder.sar(shift, shifted)
     }
 
-    /// Lowers a `new T[](len)` memory array expression.
-    fn lower_new_array(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        ty: &hir::Type<'_>,
-        args: &CallArgs<'_>,
-    ) -> ValueId {
-        if !self.is_memory_array_new_type(ty) {
-            return builder.imm_u64(0);
-        }
-
-        let len = args
-            .exprs()
-            .next()
-            .map(|arg| self.lower_expr(builder, arg))
-            .unwrap_or_else(|| builder.imm_u64(0));
-
-        let free_ptr_addr = builder.imm_u64(0x40);
-        let ptr = builder.mload(free_ptr_addr);
-        builder.mstore(ptr, len);
-
-        let word_size = builder.imm_u64(32);
-        let data_size = if matches!(
-            &ty.kind,
-            hir::TypeKind::Elementary(ElementaryType::Bytes | ElementaryType::String)
-        ) {
-            // `bytes`/`string`: the length counts bytes; the data area is the
-            // length padded up to a word.
-            let thirty_one = builder.imm_u64(31);
-            let rounded = builder.add(len, thirty_one);
-            let rounded_overflow = builder.lt(rounded, len);
-            self.emit_panic_if(builder, rounded_overflow, PanicCode::MemoryAllocationOverflow);
-            let mask = builder.not(thirty_one);
-            builder.and(rounded, mask)
-        } else {
-            // Arrays: one word per element.
-            let data_size = builder.mul(len, word_size);
-            let checked_len = builder.div(data_size, word_size);
-            let overflow = builder.eq(checked_len, len);
-            self.emit_panic_if_zero(builder, overflow, PanicCode::MemoryAllocationOverflow);
-            data_size
-        };
-        let total_size = builder.add(data_size, word_size);
-        let total_overflow = builder.lt(total_size, data_size);
-        self.emit_panic_if(builder, total_overflow, PanicCode::MemoryAllocationOverflow);
-        let new_free_ptr = builder.add(ptr, total_size);
-        let bump_overflow = builder.lt(new_free_ptr, ptr);
-        self.emit_panic_if(builder, bump_overflow, PanicCode::MemoryAllocationOverflow);
-        // Solidity caps memory at 2^64 bytes: an allocation past that limit
-        // panics (0x41) rather than running the VM out of gas on a huge size.
-        let mem_limit = builder.imm_u64(0xffff_ffff_ffff_ffff);
-        let over_limit = builder.gt(new_free_ptr, mem_limit);
-        self.emit_panic_if(builder, over_limit, PanicCode::MemoryAllocationOverflow);
-        let free_ptr_addr = builder.imm_u64(0x40);
-        builder.mstore(free_ptr_addr, new_free_ptr);
-
-        // Zero-initialize the data area: memory past the free pointer can be
-        // dirty (keccak staging fast paths write there without bumping it).
-        // `calldatacopy` from the end of calldata writes zeroes.
-        let data_ptr = builder.add(ptr, word_size);
-        let cds = builder.calldatasize();
-        builder.calldatacopy(data_ptr, cds, data_size);
-
-        ptr
-    }
-
-    fn is_memory_array_new_type(&self, ty: &hir::Type<'_>) -> bool {
-        match &ty.kind {
-            hir::TypeKind::Array(array) => array.size.is_none(),
-            hir::TypeKind::Elementary(ElementaryType::Bytes | ElementaryType::String) => true,
-            _ => false,
-        }
-    }
-
-    /// Lowers a `new Contract(args)` expression.
-    /// Supports call options like `new Contract{salt: s, value: v}(args)`.
-    fn lower_new_contract(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        ty: &hir::Type<'_>,
-        args: &CallArgs<'_>,
-        call_opts: Option<&[hir::NamedArg<'_>]>,
-    ) -> ValueId {
-        // Extract ContractId from the type
-        let contract_id = match &ty.kind {
-            hir::TypeKind::Custom(hir::ItemId::Contract(id)) => *id,
-            _ => {
-                self.gcx
-                    .dcx()
-                    .err("codegen expected a contract type for `new` expression")
-                    .span(ty.span)
-                    .emit();
-                return builder.imm_u64(0);
-            }
-        };
-
-        // Look up pre-compiled bytecode
-        let (bytecode, _segment_idx) = match self.contract_bytecodes.get(&contract_id) {
-            Some(bc) => bc.clone(),
-            None => {
-                self.gcx
-                    .dcx()
-                    .err(format!(
-                        "codegen is missing creation bytecode for `new {}`",
-                        self.gcx.hir.contract(contract_id).name
-                    ))
-                    .span(ty.span)
-                    .note("the deployed contract did not compile or was not lowered first")
-                    .emit();
-                return builder.imm_u64(0);
-            }
-        };
-
-        let bytecode_len = bytecode.len();
-
-        // Extract call options (salt, value)
-        let mut salt_opt: Option<ValueId> = None;
-        let mut value_opt: Option<ValueId> = None;
-
-        if let Some(opts) = call_opts {
-            for opt in opts {
-                let name = opt.name.name.as_str();
-                match name {
-                    "salt" => {
-                        salt_opt = Some(self.lower_expr(builder, &opt.value));
-                    }
-                    "value" => {
-                        value_opt = Some(self.lower_expr(builder, &opt.value));
-                    }
-                    _ => {
-                        // gas option is not supported for contract creation
-                    }
-                }
-            }
-        }
-
-        // Allocate memory for bytecode + constructor args from free memory pointer
-        let free_mem_ptr_slot = builder.imm_u64(0x40);
-        let mem_offset = builder.mload(free_mem_ptr_slot);
-
-        // Copy bytecode to memory using MSTORE
-        // For each 32-byte chunk of bytecode, emit an MSTORE at (mem_offset + offset)
-        for (i, chunk) in bytecode.chunks(32).enumerate() {
-            let mut padded = [0u8; 32];
-            padded[..chunk.len()].copy_from_slice(chunk);
-            let value = U256::from_be_bytes(padded);
-            let val_id = builder.imm_u256(value);
-            let chunk_offset = builder.imm_u64((i as u64) * 32);
-            let dest = builder.add(mem_offset, chunk_offset);
-            builder.mstore(dest, val_id);
-        }
-
-        // Append constructor arguments after bytecode
-        let mut args_offset = bytecode_len as u64;
-        for arg in args.exprs() {
-            let arg_val = self.lower_expr(builder, arg);
-            let arg_offset_imm = builder.imm_u64(args_offset);
-            let arg_dest = builder.add(mem_offset, arg_offset_imm);
-            builder.mstore(arg_dest, arg_val);
-            args_offset += 32; // Each arg is 32 bytes ABI encoded
-        }
-
-        // Total size = bytecode + args
-        let total_size = builder.imm_u64(args_offset);
-
-        // Update free memory pointer: new_free = mem_offset + ((total_size + 31) & ~31)
-        let thirty_one = builder.imm_u64(31);
-        let aligned_size = builder.add(total_size, thirty_one);
-        let mask = builder.imm_u256(U256::from(!31u64));
-        let aligned_size = builder.and(aligned_size, mask);
-        let new_free = builder.add(mem_offset, aligned_size);
-        builder.mstore(free_mem_ptr_slot, new_free);
-
-        // Value to send with CREATE/CREATE2 (0 for non-payable, or from value option)
-        let value = value_opt.unwrap_or_else(|| builder.imm_u64(0));
-
-        // Emit CREATE2 if salt is provided, otherwise CREATE
-        if let Some(salt) = salt_opt {
-            builder.create2(value, mem_offset, total_size, salt)
-        } else {
-            builder.create(value, mem_offset, total_size)
-        }
-    }
-
-    /// Lowers a builtin function call.
-    fn lower_builtin_call(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        builtin: Builtin,
-        args: &CallArgs<'_>,
-    ) -> ValueId {
-        match builtin {
-            Builtin::Keccak256 => {
-                let mut exprs = args.exprs();
-                if let Some(first) = exprs.next() {
-                    if let ExprKind::Lit(lit) = &first.kind
-                        && let LitKind::Str(_, bytes, _) = &lit.kind
-                    {
-                        let hash = keccak256(bytes.as_byte_str());
-                        return builder.imm_u256(U256::from_be_bytes(hash.0));
-                    }
-
-                    if let Some(packed_args) = self.abi_encode_packed_call_args(first) {
-                        return self.lower_keccak_abi_encode_packed(builder, packed_args);
-                    }
-                    if let Some(encode_args) = self.abi_encode_call_args(first) {
-                        let arg_exprs: Vec<_> = encode_args.exprs().collect();
-                        if let Some(hash) = self.lower_keccak_abi_encode(builder, &arg_exprs) {
-                            return hash;
-                        }
-                    }
-
-                    // Dynamic `bytes`/`string` (incl. `bytes(s)` of a calldata
-                    // value): hash the raw data after materializing it to memory.
-                    if let Some(hash) = self.keccak_dynamic_bytes(builder, first) {
-                        return hash;
-                    }
-                    let arg_val = self.lower_expr(builder, first);
-                    let ptr = builder.imm_u64(0);
-                    builder.mstore(ptr, arg_val);
-                    let size = builder.imm_u64(32);
-                    return builder.keccak256(ptr, size);
-                }
-                builder.imm_u64(0)
-            }
-            Builtin::Require | Builtin::Assert => {
-                let mut exprs = args.exprs();
-                if let Some(first) = exprs.next() {
-                    let cond = self.lower_expr(builder, first);
-                    let is_false = builder.iszero(cond);
-
-                    let revert_block = builder.create_block();
-                    let continue_block = builder.create_block();
-
-                    builder.branch(is_false, revert_block, continue_block);
-
-                    builder.switch_to_block(revert_block);
-                    if matches!(builtin, Builtin::Assert) {
-                        self.emit_panic_revert(builder, PanicCode::Assert);
-                    } else if let Some(message) = exprs.next() {
-                        if !self.emit_revert_payload_from_expr(builder, message) {
-                            let zero = builder.imm_u64(0);
-                            builder.revert(zero, zero);
-                        }
-                    } else {
-                        let zero = builder.imm_u64(0);
-                        builder.revert(zero, zero);
-                    }
-
-                    builder.switch_to_block(continue_block);
-                }
-                builder.imm_u64(0)
-            }
-            Builtin::Revert => {
-                let zero = builder.imm_u64(0);
-                builder.revert(zero, zero);
-                zero
-            }
-            Builtin::RevertMsg => {
-                let mut exprs = args.exprs();
-                let emitted = exprs.next().is_some_and(|message| {
-                    self.emit_revert_error_string_from_expr(builder, message)
-                });
-                let zero = builder.imm_u64(0);
-                if !emitted {
-                    builder.revert(zero, zero);
-                }
-                zero
-            }
-            Builtin::AddressBalance => {
-                let mut exprs = args.exprs();
-                if let Some(first) = exprs.next() {
-                    let addr = self.lower_expr(builder, first);
-                    return builder.balance(addr);
-                }
-                builder.imm_u64(0)
-            }
-            Builtin::AddMod | Builtin::MulMod => {
-                let mut exprs = args.exprs();
-                let Some(a) = exprs.next() else { return builder.imm_u64(0) };
-                let Some(b) = exprs.next() else { return builder.imm_u64(0) };
-                let Some(n) = exprs.next() else { return builder.imm_u64(0) };
-                let a = self.lower_expr(builder, a);
-                let b = self.lower_expr(builder, b);
-                let n = self.lower_expr(builder, n);
-                if matches!(builtin, Builtin::AddMod) {
-                    builder.addmod(a, b, n)
-                } else {
-                    builder.mulmod(a, b, n)
-                }
-            }
-            Builtin::AbiEncode => {
-                // abi.encode: a fresh `bytes memory` allocation holding the
-                // padded ABI tuple encoding of the arguments.
-                let arg_exprs: Vec<_> = args.exprs().collect();
-                if let Some(ptr) = self.lower_abi_encode_to_bytes(builder, &arg_exprs) {
-                    return ptr;
-                }
-                self.gcx
-                    .dcx()
-                    .err("codegen does not support these `abi.encode` arguments yet")
-                    .span(args.span)
-                    .emit();
-                builder.imm_u64(0)
-            }
-            Builtin::AbiEncodePacked => {
-                // abi.encodePacked: pack values tightly based on their types
-                // Returns bytes memory (length + data)
-                self.lower_abi_encode_packed(builder, args)
-            }
-            Builtin::AbiDecode => self.lower_abi_decode(builder, args),
-            Builtin::YulAdd
-            | Builtin::YulSub
-            | Builtin::YulMul
-            | Builtin::YulDiv
-            | Builtin::YulMod
-            | Builtin::YulExp
-            | Builtin::YulNot
-            | Builtin::YulAnd
-            | Builtin::YulOr
-            | Builtin::YulXor
-            | Builtin::YulShl
-            | Builtin::YulShr
-            | Builtin::YulSar
-            | Builtin::YulStop
-            | Builtin::YulSdiv
-            | Builtin::YulSmod
-            | Builtin::YulLt
-            | Builtin::YulGt
-            | Builtin::YulSlt
-            | Builtin::YulSgt
-            | Builtin::YulEq
-            | Builtin::YulIszero
-            | Builtin::YulByte
-            | Builtin::YulClz
-            | Builtin::YulAddmod
-            | Builtin::YulMulmod
-            | Builtin::YulSignextend
-            | Builtin::YulKeccak256
-            | Builtin::YulAddress
-            | Builtin::YulBalance
-            | Builtin::YulSelfbalance
-            | Builtin::YulCaller
-            | Builtin::YulCallvalue
-            | Builtin::YulCalldataload
-            | Builtin::YulCalldatasize
-            | Builtin::YulCalldatacopy
-            | Builtin::YulCodesize
-            | Builtin::YulCodecopy
-            | Builtin::YulExtcodesize
-            | Builtin::YulExtcodecopy
-            | Builtin::YulReturndatasize
-            | Builtin::YulReturndatacopy
-            | Builtin::YulExtcodehash
-            | Builtin::YulMload
-            | Builtin::YulMstore
-            | Builtin::YulMstore8
-            | Builtin::YulSload
-            | Builtin::YulSstore
-            | Builtin::YulTload
-            | Builtin::YulTstore
-            | Builtin::YulMsize
-            | Builtin::YulGas
-            | Builtin::YulLog0
-            | Builtin::YulLog1
-            | Builtin::YulLog2
-            | Builtin::YulLog3
-            | Builtin::YulLog4
-            | Builtin::YulCreate
-            | Builtin::YulCreate2
-            | Builtin::YulCall
-            | Builtin::YulCallcode
-            | Builtin::YulDelegatecall
-            | Builtin::YulStaticcall
-            | Builtin::YulExtcall
-            | Builtin::YulExtdelegatecall
-            | Builtin::YulExtstaticcall
-            | Builtin::YulReturn
-            | Builtin::YulRevert
-            | Builtin::YulSelfdestruct
-            | Builtin::YulInvalid
-            | Builtin::YulChainid
-            | Builtin::YulBasefee
-            | Builtin::YulBlobbasefee
-            | Builtin::YulBlobhash
-            | Builtin::YulCoinbase
-            | Builtin::YulDifficulty
-            | Builtin::YulPrevrandao
-            | Builtin::YulGaslimit
-            | Builtin::YulNumber
-            | Builtin::YulTimestamp
-            | Builtin::YulGasprice
-            | Builtin::YulOrigin
-            | Builtin::YulBlockhash
-            | Builtin::YulPop
-            | Builtin::YulMcopy => self.lower_yul_builtin_call(builder, builtin, args),
-            _ => builder.imm_u64(0),
-        }
-    }
-
-    fn lower_yul_builtin_call(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        builtin: Builtin,
-        args: &CallArgs<'_>,
-    ) -> ValueId {
-        let arg_vals: Vec<ValueId> =
-            args.exprs().map(|arg| self.lower_expr(builder, arg)).collect();
-        if let Some(expected) = Self::yul_builtin_arity(builtin)
-            && arg_vals.len() != expected
-        {
-            self.gcx
-                .dcx()
-                .err(format!(
-                    "wrong number of arguments for Yul builtin `{}`: expected {}, found {}",
-                    builtin.name(),
-                    expected,
-                    arg_vals.len()
-                ))
-                .span(args.span)
-                .emit();
-            return builder.imm_u64(0);
-        }
-
-        match builtin {
-            Builtin::YulAdd => builder.add(arg_vals[0], arg_vals[1]),
-            Builtin::YulSub => builder.sub(arg_vals[0], arg_vals[1]),
-            Builtin::YulMul => builder.mul(arg_vals[0], arg_vals[1]),
-            Builtin::YulDiv => builder.div(arg_vals[0], arg_vals[1]),
-            Builtin::YulSdiv => builder.sdiv(arg_vals[0], arg_vals[1]),
-            Builtin::YulMod => builder.mod_(arg_vals[0], arg_vals[1]),
-            Builtin::YulSmod => builder.smod(arg_vals[0], arg_vals[1]),
-            Builtin::YulAddmod => builder.addmod(arg_vals[0], arg_vals[1], arg_vals[2]),
-            Builtin::YulMulmod => builder.mulmod(arg_vals[0], arg_vals[1], arg_vals[2]),
-            Builtin::YulExp => builder.exp(arg_vals[0], arg_vals[1]),
-            Builtin::YulSignextend => builder.signextend(arg_vals[0], arg_vals[1]),
-            Builtin::YulAnd => builder.and(arg_vals[0], arg_vals[1]),
-            Builtin::YulOr => builder.or(arg_vals[0], arg_vals[1]),
-            Builtin::YulXor => builder.xor(arg_vals[0], arg_vals[1]),
-            Builtin::YulNot => builder.not(arg_vals[0]),
-            Builtin::YulByte => builder.byte(arg_vals[0], arg_vals[1]),
-            Builtin::YulShl => builder.shl(arg_vals[0], arg_vals[1]),
-            Builtin::YulShr => builder.shr(arg_vals[0], arg_vals[1]),
-            Builtin::YulSar => builder.sar(arg_vals[0], arg_vals[1]),
-            Builtin::YulLt => builder.lt(arg_vals[0], arg_vals[1]),
-            Builtin::YulGt => builder.gt(arg_vals[0], arg_vals[1]),
-            Builtin::YulSlt => builder.slt(arg_vals[0], arg_vals[1]),
-            Builtin::YulSgt => builder.sgt(arg_vals[0], arg_vals[1]),
-            Builtin::YulEq => builder.eq(arg_vals[0], arg_vals[1]),
-            Builtin::YulIszero => builder.iszero(arg_vals[0]),
-            Builtin::YulMload => builder.mload(arg_vals[0]),
-            Builtin::YulMstore => {
-                builder.mstore(arg_vals[0], arg_vals[1]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulMstore8 => {
-                builder.mstore8(arg_vals[0], arg_vals[1]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulMsize => builder.msize(),
-            Builtin::YulMcopy => {
-                self.mcopy(builder, arg_vals[0], arg_vals[1], arg_vals[2], Some(args.span));
-                builder.imm_u64(0)
-            }
-            Builtin::YulSload => builder.sload(arg_vals[0]),
-            Builtin::YulSstore => {
-                builder.sstore(arg_vals[0], arg_vals[1]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulTload => builder.tload(arg_vals[0]),
-            Builtin::YulTstore => {
-                builder.tstore(arg_vals[0], arg_vals[1]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulCalldataload => builder.calldataload(arg_vals[0]),
-            Builtin::YulCalldatasize => builder.calldatasize(),
-            Builtin::YulCalldatacopy => {
-                builder.calldatacopy(arg_vals[0], arg_vals[1], arg_vals[2]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulCodesize => builder.codesize(),
-            Builtin::YulCodecopy => {
-                builder.codecopy(arg_vals[0], arg_vals[1], arg_vals[2]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulExtcodesize => builder.extcodesize(arg_vals[0]),
-            Builtin::YulExtcodecopy => {
-                builder.extcodecopy(arg_vals[0], arg_vals[1], arg_vals[2], arg_vals[3]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulExtcodehash => builder.extcodehash(arg_vals[0]),
-            Builtin::YulReturndatasize => builder.returndatasize(),
-            Builtin::YulReturndatacopy => {
-                builder.returndatacopy(arg_vals[0], arg_vals[1], arg_vals[2]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulAddress => builder.address(),
-            Builtin::YulBalance => builder.balance(arg_vals[0]),
-            Builtin::YulSelfbalance => builder.selfbalance(),
-            Builtin::YulCaller => builder.caller(),
-            Builtin::YulCallvalue => builder.callvalue(),
-            Builtin::YulOrigin => builder.origin(),
-            Builtin::YulGasprice => builder.gasprice(),
-            Builtin::YulBlockhash => builder.blockhash(arg_vals[0]),
-            Builtin::YulCoinbase => builder.coinbase(),
-            Builtin::YulTimestamp => builder.timestamp(),
-            Builtin::YulNumber => builder.number(),
-            Builtin::YulDifficulty | Builtin::YulPrevrandao => builder.prevrandao(),
-            Builtin::YulGaslimit => builder.gaslimit(),
-            Builtin::YulChainid => builder.chainid(),
-            Builtin::YulGas => builder.gas(),
-            Builtin::YulBasefee => builder.basefee(),
-            Builtin::YulBlobbasefee => builder.blobbasefee(),
-            Builtin::YulBlobhash => builder.blobhash(arg_vals[0]),
-            Builtin::YulKeccak256 => builder.keccak256(arg_vals[0], arg_vals[1]),
-            Builtin::YulCall => builder.call(
-                arg_vals[0],
-                arg_vals[1],
-                arg_vals[2],
-                arg_vals[3],
-                arg_vals[4],
-                arg_vals[5],
-                arg_vals[6],
-            ),
-            Builtin::YulStaticcall => builder.staticcall(
-                arg_vals[0],
-                arg_vals[1],
-                arg_vals[2],
-                arg_vals[3],
-                arg_vals[4],
-                arg_vals[5],
-            ),
-            Builtin::YulDelegatecall => builder.delegatecall(
-                arg_vals[0],
-                arg_vals[1],
-                arg_vals[2],
-                arg_vals[3],
-                arg_vals[4],
-                arg_vals[5],
-            ),
-            Builtin::YulCreate => builder.create(arg_vals[0], arg_vals[1], arg_vals[2]),
-            Builtin::YulCreate2 => {
-                builder.create2(arg_vals[0], arg_vals[1], arg_vals[2], arg_vals[3])
-            }
-            Builtin::YulLog0 => {
-                builder.log0(arg_vals[0], arg_vals[1]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulLog1 => {
-                builder.log1(arg_vals[0], arg_vals[1], arg_vals[2]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulLog2 => {
-                builder.log2(arg_vals[0], arg_vals[1], arg_vals[2], arg_vals[3]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulLog3 => {
-                builder.log3(arg_vals[0], arg_vals[1], arg_vals[2], arg_vals[3], arg_vals[4]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulLog4 => {
-                builder.log4(
-                    arg_vals[0],
-                    arg_vals[1],
-                    arg_vals[2],
-                    arg_vals[3],
-                    arg_vals[4],
-                    arg_vals[5],
-                );
-                builder.imm_u64(0)
-            }
-            Builtin::YulRevert => {
-                builder.revert(arg_vals[0], arg_vals[1]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulStop => {
-                builder.stop();
-                builder.imm_u64(0)
-            }
-            Builtin::YulInvalid => {
-                builder.invalid();
-                builder.imm_u64(0)
-            }
-            Builtin::YulSelfdestruct => {
-                builder.selfdestruct(arg_vals[0]);
-                builder.imm_u64(0)
-            }
-            Builtin::YulPop => builder.imm_u64(0),
-            Builtin::YulClz
-            | Builtin::YulCallcode
-            | Builtin::YulExtcall
-            | Builtin::YulExtdelegatecall
-            | Builtin::YulExtstaticcall
-            | Builtin::YulReturn => self.unsupported_yul_builtin(builder, builtin, args.span),
-            _ => unreachable!("non-Yul builtin passed to Yul lowering"),
-        }
-    }
-
-    fn yul_builtin_arity(builtin: Builtin) -> Option<usize> {
-        Some(match builtin {
-            Builtin::YulStop
-            | Builtin::YulAddress
-            | Builtin::YulSelfbalance
-            | Builtin::YulCaller
-            | Builtin::YulCallvalue
-            | Builtin::YulCalldatasize
-            | Builtin::YulCodesize
-            | Builtin::YulReturndatasize
-            | Builtin::YulMsize
-            | Builtin::YulGas
-            | Builtin::YulInvalid
-            | Builtin::YulChainid
-            | Builtin::YulBasefee
-            | Builtin::YulBlobbasefee
-            | Builtin::YulCoinbase
-            | Builtin::YulDifficulty
-            | Builtin::YulPrevrandao
-            | Builtin::YulGaslimit
-            | Builtin::YulNumber
-            | Builtin::YulTimestamp
-            | Builtin::YulGasprice
-            | Builtin::YulOrigin => 0,
-            Builtin::YulNot
-            | Builtin::YulIszero
-            | Builtin::YulClz
-            | Builtin::YulBalance
-            | Builtin::YulCalldataload
-            | Builtin::YulExtcodesize
-            | Builtin::YulExtcodehash
-            | Builtin::YulMload
-            | Builtin::YulSload
-            | Builtin::YulTload
-            | Builtin::YulBlobhash
-            | Builtin::YulBlockhash
-            | Builtin::YulPop
-            | Builtin::YulSelfdestruct => 1,
-            Builtin::YulAdd
-            | Builtin::YulSub
-            | Builtin::YulMul
-            | Builtin::YulDiv
-            | Builtin::YulMod
-            | Builtin::YulExp
-            | Builtin::YulAnd
-            | Builtin::YulOr
-            | Builtin::YulXor
-            | Builtin::YulShl
-            | Builtin::YulShr
-            | Builtin::YulSar
-            | Builtin::YulSdiv
-            | Builtin::YulSmod
-            | Builtin::YulLt
-            | Builtin::YulGt
-            | Builtin::YulSlt
-            | Builtin::YulSgt
-            | Builtin::YulEq
-            | Builtin::YulByte
-            | Builtin::YulSignextend
-            | Builtin::YulKeccak256
-            | Builtin::YulMstore
-            | Builtin::YulMstore8
-            | Builtin::YulSstore
-            | Builtin::YulTstore
-            | Builtin::YulLog0
-            | Builtin::YulReturn
-            | Builtin::YulRevert => 2,
-            Builtin::YulAddmod
-            | Builtin::YulMulmod
-            | Builtin::YulCalldatacopy
-            | Builtin::YulCodecopy
-            | Builtin::YulReturndatacopy
-            | Builtin::YulMcopy
-            | Builtin::YulLog1
-            | Builtin::YulCreate
-            | Builtin::YulExtdelegatecall
-            | Builtin::YulExtstaticcall => 3,
-            Builtin::YulExtcodecopy
-            | Builtin::YulLog2
-            | Builtin::YulCreate2
-            | Builtin::YulExtcall => 4,
-            Builtin::YulLog3 => 5,
-            Builtin::YulDelegatecall | Builtin::YulStaticcall | Builtin::YulLog4 => 6,
-            Builtin::YulCall | Builtin::YulCallcode => 7,
-            _ => return None,
-        })
-    }
-
-    fn unsupported_yul_builtin(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        builtin: Builtin,
-        span: Span,
-    ) -> ValueId {
-        self.gcx
-            .dcx()
-            .err(format!("unsupported Yul builtin `{}`", builtin.name()))
-            .span(span)
-            .emit();
-        builder.imm_u64(0)
-    }
-
-    /// Lowers a member function call (e.g., counter.increment()).
-    fn lower_member_call_with_opts(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        base: &hir::Expr<'_>,
-        member: Ident,
-        args: &CallArgs<'_>,
-        call_opts: Option<&[hir::NamedArg<'_>]>,
-    ) -> ValueId {
-        // Handle builtin member calls: abi.encode(), abi.encodePacked(), etc.
-        if let ExprKind::Ident(res_slice) = &base.kind
-            && let Some(hir::Res::Builtin(base_builtin)) = res_slice.first()
-            && let Some(member_builtin) = self.resolve_builtin_member(*base_builtin, member.name)
-        {
-            return self.lower_builtin_call(builder, member_builtin, args);
-        }
-
-        // Handle library function calls: Library.func(args)
-        // The base is an Ident resolving to a ContractId for a library
-        if let ExprKind::Ident(res_slice) = &base.kind
-            && let Some(hir::Res::Item(hir::ItemId::Contract(contract_id))) = res_slice.first()
-        {
-            let contract = self.gcx.hir.contract(*contract_id);
-
-            // Check if member is a struct type defined in this contract: Contract.StructType(args)
-            for &item_id in contract.items {
-                if let hir::ItemId::Struct(struct_id) = item_id {
-                    let strukt = self.gcx.hir.strukt(struct_id);
-                    if strukt.name.name == member.name {
-                        return self.lower_struct_constructor(builder, struct_id, args);
-                    }
-                }
-            }
-
-            if contract.kind.is_library() {
-                // Find the library function by name, matching argument count for overloads
-                let arg_count = args.exprs().count();
-                if let Some(func_id) =
-                    self.find_library_function(*contract_id, member.name, arg_count)
-                {
-                    return self.lower_library_call(builder, func_id, args, None);
-                }
-            }
-        }
-
-        // Handle address payable transfer/send builtins
-        // Only treat as address builtins if base is NOT a contract type
-        let member_name = member.name.as_str();
-        if (member_name == "transfer" || member_name == "send") && !self.is_contract_type_expr(base)
-        {
-            // payable(addr).transfer(amount) or payable(addr).send(amount)
-            // CALL(2300, addr, amount, 0, 0, 0, 0)
-            let addr = self.lower_expr(builder, base);
-            let mut exprs = args.exprs();
-            let amount = if let Some(first) = exprs.next() {
-                self.lower_expr(builder, first)
-            } else {
-                builder.imm_u64(0)
-            };
-
-            // transfer/send uses 2300 gas stipend
-            let gas_stipend = builder.imm_u64(2300);
-            // Create fresh zero values for each CALL argument to avoid stack issues
-            let zero_args_offset = builder.imm_u64(0);
-            let zero_args_size = builder.imm_u64(0);
-            let zero_ret_offset = builder.imm_u64(0);
-            let zero_ret_size = builder.imm_u64(0);
-
-            // CALL(gas, addr, value, argsOffset, argsSize, retOffset, retSize)
-            let success = builder.call(
-                gas_stipend,
-                addr,
-                amount,
-                zero_args_offset,
-                zero_args_size,
-                zero_ret_offset,
-                zero_ret_size,
-            );
-
-            if member_name == "transfer" {
-                // transfer reverts on failure
-                let is_failure = builder.iszero(success);
-                let revert_block = builder.create_block();
-                let continue_block = builder.create_block();
-                builder.branch(is_failure, revert_block, continue_block);
-                builder.switch_to_block(revert_block);
-                let revert_offset = builder.imm_u64(0);
-                let revert_size = builder.imm_u64(0);
-                builder.revert(revert_offset, revert_size);
-                builder.switch_to_block(continue_block);
-                return builder.imm_u64(0);
-            }
-            // send returns success bool
-            return success;
-        }
-
-        // Handle low-level call/staticcall/delegatecall
-        // addr.call{value: X}(data) returns (bool success, bytes memory returndata)
-        // addr.staticcall(data) returns (bool success, bytes memory returndata)
-        // addr.delegatecall(data) returns (bool success, bytes memory returndata)
-        if member_name == "call" || member_name == "staticcall" || member_name == "delegatecall" {
-            let addr = self.lower_expr(builder, base);
-
-            // Get the calldata bytes argument.
-            let mut exprs = args.exprs();
-            let (calldata_offset, calldata_size) = if let Some(data_arg) = exprs.next() {
-                // Supported inputs are literals and ABI encode calls. Other
-                // bytes expressions panic in `lower_bytes_arg_to_memory`.
-                self.lower_bytes_arg_to_memory(builder, data_arg)
-            } else {
-                // No argument means empty calldata
-                (builder.imm_u64(0), builder.imm_u64(0))
-            };
-
-            // Gas: use all available gas
-            let gas = builder.gas();
-
-            // Value: extract from call options {value: X} or default to 0
-            let value = if member_name == "call" {
-                self.extract_call_value(builder, call_opts)
-            } else {
-                // staticcall and delegatecall don't transfer value
-                builder.imm_u64(0)
-            };
-
-            // This lowering models only the success flag. Solidity's second
-            // `bytes` result is rejected by `lower_multi_var_decl` until the
-            // compiler materializes returndata bytes.
-            let ret_offset = builder.imm_u64(0);
-            let ret_size = builder.imm_u64(0);
-
-            // Emit the appropriate CALL/STATICCALL/DELEGATECALL instruction
-            let success = match member_name {
-                "call" => builder.call(
-                    gas,
-                    addr,
-                    value,
-                    calldata_offset,
-                    calldata_size,
-                    ret_offset,
-                    ret_size,
-                ),
-                "staticcall" => builder.staticcall(
-                    gas,
-                    addr,
-                    calldata_offset,
-                    calldata_size,
-                    ret_offset,
-                    ret_size,
-                ),
-                "delegatecall" => builder.delegatecall(
-                    gas,
-                    addr,
-                    calldata_offset,
-                    calldata_size,
-                    ret_offset,
-                    ret_size,
-                ),
-                _ => unreachable!(),
-            };
-
-            // Low-level calls return `(bool, bytes)`, but this expression path
-            // exposes only the first value. `lower_multi_var_decl` copies the
-            // returndata bytes out of the return buffer when they are bound.
-            return success;
-        }
-
-        // Handle storage `bytes`/`string` methods before the generic member
-        // call path. Their storage layout is Solidity's packed short/long
-        // bytes form, not the generic dynamic-array layout.
-        if self.is_storage_bytes_expr(base)
-            && matches!(member_name, "push" | "pop")
-            && let Some(slot) = self.lower_lvalue_slot(builder, base)
-        {
-            return self.lower_storage_bytes_method_call(builder, slot, member_name, args);
-        }
-
-        // Handle dynamic array methods (push, pop)
-        if let Some((var_id, slot)) = self.get_dyn_array_base_slot(base) {
-            return self.lower_array_method_call(builder, var_id, slot, member_name, args);
-        }
-
-        // Handle `using X for Y` library calls: x.method(args) -> Library.method(x, args)
-        if let Some(func_id) = self.resolve_using_directive_call(base, member.name) {
-            let bound_arg = self.lower_expr(builder, base);
-            return self.lower_library_call(builder, func_id, args, Some(bound_arg));
-        }
-
-        // Look up the function being called to get its selector and return count
-        let selector = self.compute_member_selector(base, member);
-        let num_returns = self.get_member_function_return_count(base, member);
-
-        // Collect argument info: for structs we need the field count, for scalars just 1 slot
-        let arg_infos: Vec<_> = args
-            .exprs()
-            .map(|arg| {
-                let struct_info = self.get_expr_struct_info(arg);
-                (arg, struct_info)
-            })
-            .collect();
-
-        // Calculate calldata size: 4 bytes selector + sum of all argument slots
-        let total_arg_slots: usize =
-            arg_infos.iter().map(|(_, info)| info.map(|(_, n)| n).unwrap_or(1)).sum();
-        let calldata_size_bytes = 4 + total_arg_slots * 32;
-
-        // IMPORTANT: Evaluate all arguments FIRST before writing to memory.
-        // For structs, lower_expr returns the memory pointer.
-        let arg_vals: Vec<ValueId> =
-            arg_infos.iter().map(|(arg, _)| self.lower_expr(builder, arg)).collect();
-
-        // Evaluate the address and spill it to scratch memory at 0x00.
-        // This ensures it survives all the MSTORE operations for calldata setup.
-        // We reload it right before the CALL.
-        let addr_expr = self.lower_expr(builder, base);
-        let scratch_addr = builder.imm_u64(0x00);
-        builder.mstore(scratch_addr, addr_expr);
-
-        // Allocate calldata from the free memory pointer (like solc does).
-        // This avoids clobbering the free memory pointer at 0x40 when encoding
-        // calldata with 2+ arguments (which would span 0x04-0x43+).
-        let free_ptr_addr = builder.imm_u64(0x40);
-        let calldata_start = builder.mload(free_ptr_addr);
-
-        // Store calldata_start to scratch memory at 0x20.
-        // We need to reload it right before the CALL because:
-        // 1. The scheduler may lose track of this value after many MSTOREs
-        // 2. For struct returns, we update the free memory pointer, so reading 0x40 again would be
-        //    wrong
-        let scratch_calldata = builder.imm_u64(0x20);
-        builder.mstore(scratch_calldata, calldata_start);
-
-        // Write the selector at calldata_start (left-aligned in 32-byte word)
-        let selector_word = U256::from(selector) << 224;
-        let selector_val = builder.imm_u256(selector_word);
-        builder.mstore(calldata_start, selector_val);
-
-        // Write arguments after selector
-        // For struct arguments, we need to load each field from memory and write them
-        let mut arg_offset = 4u64;
-        for (i, arg_val) in arg_vals.iter().enumerate() {
-            let struct_info = &arg_infos[i].1;
-
-            if let Some((_, field_count)) = struct_info {
-                // Struct argument: load each field from memory and write to calldata
-                for field_idx in 0..*field_count {
-                    let field_mem_offset = (field_idx as u64) * 32;
-                    let field_val = if field_mem_offset == 0 {
-                        builder.mload(*arg_val)
-                    } else {
-                        let field_offset_val = builder.imm_u64(field_mem_offset);
-                        let field_addr = builder.add(*arg_val, field_offset_val);
-                        builder.mload(field_addr)
-                    };
-
-                    let offset_val = builder.imm_u64(arg_offset);
-                    let write_addr = builder.add(calldata_start, offset_val);
-                    builder.mstore(write_addr, field_val);
-                    arg_offset += 32;
-                }
-            } else {
-                // Scalar argument: write directly
-                let offset_val = builder.imm_u64(arg_offset);
-                let write_addr = builder.add(calldata_start, offset_val);
-                builder.mstore(write_addr, *arg_val);
-                arg_offset += 32;
-            }
-        }
-
-        // Check if the return type is a struct - need special handling for return data
-        let struct_return_info = self.get_member_function_struct_return(base, member);
-
-        // Determine where to store return data and whether it's a struct
-        let (ret_offset, ret_size, struct_ptr_opt) =
-            if let Some((_struct_id, field_count)) = struct_return_info {
-                // For struct returns: allocate space after calldata for the return value
-                let struct_size = (field_count as u64) * 32;
-                let calldata_end_offset = builder.imm_u64(calldata_size_bytes as u64);
-                let struct_ptr = builder.add(calldata_start, calldata_end_offset);
-
-                // Update free memory pointer past the struct
-                let struct_size_val = builder.imm_u64(struct_size);
-                let new_free_ptr = builder.add(struct_ptr, struct_size_val);
-                builder.mstore(free_ptr_addr, new_free_ptr);
-
-                let ret_size = builder.imm_u64(struct_size);
-                (struct_ptr, ret_size, Some(struct_ptr))
-            } else {
-                // For non-struct returns: use scratch space at offset 0
-                // (safe because we're done with calldata after the CALL)
-                let ret_offset = builder.imm_u64(0);
-                let ret_size = builder.imm_u64((num_returns * 32) as u64);
-                (ret_offset, ret_size, None)
-            };
-
-        // Total calldata size = 4 (selector) + 32 * num_args
-        let calldata_size = builder.imm_u64(calldata_size_bytes as u64);
-
-        // Value: extract from call options {value: X} or default to 0
-        let value = self.extract_call_value(builder, call_opts);
-
-        // Reload the address from scratch memory (0x00) where we stored it earlier.
-        // This avoids stack depth issues after all the MSTORE operations.
-        let scratch_addr_reload = builder.imm_u64(0x00);
-        let addr = builder.mload(scratch_addr_reload);
-
-        // Gas: use all available gas (must be right before CALL to be on top of stack)
-        let gas = builder.gas();
-
-        // Reload calldata_start from scratch memory at 0x20.
-        // Cannot re-read from 0x40 because struct return handling may have updated it.
-        let scratch_calldata_reload = builder.imm_u64(0x20);
-        let calldata_start_reload = builder.mload(scratch_calldata_reload);
-
-        // Emit the CALL instruction
-        let _success = builder.call(
-            gas,
-            addr,
-            value,
-            calldata_start_reload,
-            calldata_size,
-            ret_offset,
-            ret_size,
-        );
-
-        // For struct returns, the data is already in the right place (at struct_ptr).
-        // Just return the pointer.
-        if let Some(struct_ptr) = struct_ptr_opt {
-            return struct_ptr;
-        }
-
-        // Load first return value from memory
-        // Note: for multi-return calls, lower_multi_var_decl will read additional values
-        // from memory at offsets 32, 64, etc.
-        builder.mload(ret_offset)
-    }
-
-    /// Extracts the `value` from call options `{value: X}`, or returns 0 if not present.
-    pub(super) fn extract_call_value(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        call_opts: Option<&[hir::NamedArg<'_>]>,
-    ) -> ValueId {
-        if let Some(opts) = call_opts {
-            for opt in opts {
-                if opt.name.name.as_str() == "value" {
-                    return self.lower_expr(builder, &opt.value);
-                }
-            }
-        }
-        builder.imm_u64(0)
-    }
-
     /// Checks if an expression is a mapping state variable and returns its var_id and storage slot.
     fn get_mapping_base_slot(&self, expr: &hir::Expr<'_>) -> Option<(hir::VariableId, u64)> {
         if let ExprKind::Ident(res_slice) = &expr.kind
@@ -2635,7 +1359,10 @@ impl<'gcx> Lowerer<'gcx> {
 
     /// Checks if an expression is a dynamic array state variable and returns its var_id and
     /// storage slot.
-    fn get_dyn_array_base_slot(&self, expr: &hir::Expr<'_>) -> Option<(hir::VariableId, u64)> {
+    pub(super) fn get_dyn_array_base_slot(
+        &self,
+        expr: &hir::Expr<'_>,
+    ) -> Option<(hir::VariableId, u64)> {
         if let ExprKind::Ident(res_slice) = &expr.kind
             && let Some(hir::Res::Item(hir::ItemId::Variable(var_id))) = res_slice.first()
         {
@@ -2787,7 +1514,7 @@ impl<'gcx> Lowerer<'gcx> {
 
     /// Lowers a struct constructor call (e.g., Point(10, 20)).
     /// Allocates memory for the struct and stores each field value.
-    fn lower_struct_constructor(
+    pub(super) fn lower_struct_constructor(
         &mut self,
         builder: &mut FunctionBuilder<'_>,
         struct_id: hir::StructId,
@@ -2847,7 +1574,7 @@ impl<'gcx> Lowerer<'gcx> {
     ///
     /// Like solc, a word that is not a clean value of `T` reverts with empty
     /// returndata instead of being silently truncated.
-    fn lower_abi_decode(
+    pub(super) fn lower_abi_decode(
         &mut self,
         builder: &mut FunctionBuilder<'_>,
         args: &CallArgs<'_>,
@@ -3559,7 +2286,7 @@ impl<'gcx> Lowerer<'gcx> {
     /// For dynamic arrays:
     /// - Length is stored at the base slot
     /// - Elements are stored at keccak256(slot) + index
-    fn lower_array_method_call(
+    pub(super) fn lower_array_method_call(
         &mut self,
         builder: &mut FunctionBuilder<'_>,
         _var_id: hir::VariableId,
@@ -4001,1005 +2728,5 @@ impl<'gcx> Lowerer<'gcx> {
             return false;
         }
         Self::is_dynamic_mapping_key(&var.ty.kind)
-    }
-
-    /// Computes the function selector for a member call.
-    pub(super) fn compute_member_selector(&self, base: &hir::Expr<'_>, member: Ident) -> u32 {
-        // Try to get the type of the base expression and find the function
-        // For contract types, we look up the function in the contract's interface
-
-        // Helper to look up selector from a contract, including inherited functions.
-        // Searches through the linearized inheritance chain.
-        let lookup_in_contract = |contract_id: hir::ContractId| -> Option<u32> {
-            let contract = self.gcx.hir.contract(contract_id);
-            // Search through the inheritance chain (linearized_bases includes self at index 0)
-            for &base_id in contract.linearized_bases.iter() {
-                let base_contract = self.gcx.hir.contract(base_id);
-                for func_id in base_contract.all_functions() {
-                    let func = self.gcx.hir.function(func_id);
-                    if func.name.is_some_and(|n| n.name == member.name) {
-                        let selector = self.gcx.function_selector(func_id);
-                        return Some(u32::from_be_bytes(selector.0));
-                    }
-                }
-            }
-            None
-        };
-
-        // Case 1: base is an identifier (variable with contract type)
-        if let ExprKind::Ident(res_slice) = &base.kind
-            && let Some(hir::Res::Item(hir::ItemId::Variable(var_id))) = res_slice.first()
-        {
-            let var = self.gcx.hir.variable(*var_id);
-            let ty = self.gcx.type_of_hir_ty(&var.ty);
-            if let solar_sema::ty::TyKind::Contract(contract_id) = ty.kind
-                && let Some(sel) = lookup_in_contract(contract_id)
-            {
-                return sel;
-            }
-        }
-
-        // Case 2: base is a type conversion call like ICallee(addr)
-        // The call's callee is an Ident resolving to a Contract/Interface
-        if let ExprKind::Call(callee, _args, _named) = &base.kind
-            && let ExprKind::Ident(res_slice) = &callee.kind
-            && let Some(hir::Res::Item(hir::ItemId::Contract(contract_id))) = res_slice.first()
-            && let Some(sel) = lookup_in_contract(*contract_id)
-        {
-            return sel;
-        }
-
-        // Case 2b: base is the contract/interface name itself, e.g.
-        // `IERC20Minimal.transfer.selector`.
-        if let ExprKind::Ident(res_slice) = &base.kind
-            && let Some(hir::Res::Item(hir::ItemId::Contract(contract_id))) = res_slice.first()
-            && let Some(sel) = lookup_in_contract(*contract_id)
-        {
-            return sel;
-        }
-
-        // Case 3: base is `this` (Builtin::This)
-        if let ExprKind::Ident(res_slice) = &base.kind
-            && let Some(hir::Res::Builtin(Builtin::This)) = res_slice.first()
-            && let Some(contract_id) = self.current_contract_id
-            && let Some(sel) = lookup_in_contract(contract_id)
-        {
-            return sel;
-        }
-
-        // Fallback: compute selector from member name
-        // This is a simplified version - proper implementation would use full signature
-        let sig = format!("{}()", member.name);
-        let hash = alloy_primitives::keccak256(sig.as_bytes());
-        u32::from_be_bytes(hash[..4].try_into().unwrap())
-    }
-
-    /// Gets the number of return values for a member function call.
-    pub(super) fn get_member_function_return_count(
-        &self,
-        base: &hir::Expr<'_>,
-        member: Ident,
-    ) -> usize {
-        // Helper to count the number of 32-byte slots a return type occupies.
-        // Structs are expanded to their number of fields.
-        let count_return_slots = |returns: &[hir::VariableId]| -> usize {
-            let mut total = 0;
-            for &var_id in returns {
-                let var = self.gcx.hir.variable(var_id);
-                if let hir::TypeKind::Custom(hir::ItemId::Struct(struct_id)) = &var.ty.kind {
-                    // Struct: count its fields
-                    let strukt = self.gcx.hir.strukt(*struct_id);
-                    total += strukt.fields.len();
-                } else {
-                    // Non-struct: 1 slot
-                    total += 1;
-                }
-            }
-            total.max(1)
-        };
-
-        // Helper to look up return count from a contract, including inherited functions.
-        // Searches through the linearized inheritance chain.
-        let lookup_in_contract = |contract_id: hir::ContractId| -> Option<usize> {
-            let contract = self.gcx.hir.contract(contract_id);
-            // Search through the inheritance chain (linearized_bases includes self at index 0)
-            for &base_id in contract.linearized_bases.iter() {
-                let base_contract = self.gcx.hir.contract(base_id);
-                for func_id in base_contract.all_functions() {
-                    let func = self.gcx.hir.function(func_id);
-                    if func.name.is_some_and(|n| n.name == member.name) {
-                        return Some(count_return_slots(func.returns));
-                    }
-                }
-            }
-            None
-        };
-
-        // Case 1: base is an identifier (variable with contract type)
-        if let ExprKind::Ident(res_slice) = &base.kind
-            && let Some(hir::Res::Item(hir::ItemId::Variable(var_id))) = res_slice.first()
-        {
-            let var = self.gcx.hir.variable(*var_id);
-            let ty = self.gcx.type_of_hir_ty(&var.ty);
-            if let solar_sema::ty::TyKind::Contract(contract_id) = ty.kind
-                && let Some(count) = lookup_in_contract(contract_id)
-            {
-                return count;
-            }
-        }
-
-        // Case 2: base is a type conversion call like ICallee(addr)
-        if let ExprKind::Call(callee, _args, _named) = &base.kind
-            && let ExprKind::Ident(res_slice) = &callee.kind
-            && let Some(hir::Res::Item(hir::ItemId::Contract(contract_id))) = res_slice.first()
-            && let Some(count) = lookup_in_contract(*contract_id)
-        {
-            return count;
-        }
-
-        // Case 3: base is `this` (Builtin::This)
-        if let ExprKind::Ident(res_slice) = &base.kind
-            && let Some(hir::Res::Builtin(Builtin::This)) = res_slice.first()
-        {
-            // Look up the function in the current contract
-            // We need to find it through the module's functions
-            // Search all known contracts because `this` carries the current
-            // contract value rather than a specific function declaration.
-            for contract_id in self.gcx.hir.contract_ids() {
-                if let Some(count) = lookup_in_contract(contract_id) {
-                    return count;
-                }
-            }
-        }
-
-        // Unknown member calls are treated as single-value calls.
-        1
-    }
-
-    /// Gets struct return info for a member function call.
-    /// Returns Some((struct_id, field_count)) if the function returns a single struct.
-    pub(super) fn get_member_function_struct_return(
-        &self,
-        base: &hir::Expr<'_>,
-        member: Ident,
-    ) -> Option<(hir::StructId, usize)> {
-        // Helper to check if the function returns a single struct and get its info.
-        let check_struct_return = |returns: &[hir::VariableId]| -> Option<(hir::StructId, usize)> {
-            if returns.len() == 1 {
-                let var = self.gcx.hir.variable(returns[0]);
-                if let hir::TypeKind::Custom(hir::ItemId::Struct(struct_id)) = &var.ty.kind {
-                    let strukt = self.gcx.hir.strukt(*struct_id);
-                    return Some((*struct_id, strukt.fields.len()));
-                }
-            }
-            None
-        };
-
-        // Helper to look up struct return from a contract, including inherited functions.
-        let lookup_in_contract = |contract_id: hir::ContractId| -> Option<(hir::StructId, usize)> {
-            let contract = self.gcx.hir.contract(contract_id);
-            for &base_id in contract.linearized_bases.iter() {
-                let base_contract = self.gcx.hir.contract(base_id);
-                for func_id in base_contract.all_functions() {
-                    let func = self.gcx.hir.function(func_id);
-                    if func.name.is_some_and(|n| n.name == member.name) {
-                        return check_struct_return(func.returns);
-                    }
-                }
-            }
-            None
-        };
-
-        // Case 1: base is an identifier (variable with contract type)
-        if let ExprKind::Ident(res_slice) = &base.kind
-            && let Some(hir::Res::Item(hir::ItemId::Variable(var_id))) = res_slice.first()
-        {
-            let var = self.gcx.hir.variable(*var_id);
-            let ty = self.gcx.type_of_hir_ty(&var.ty);
-            if let solar_sema::ty::TyKind::Contract(contract_id) = ty.kind
-                && let Some(info) = lookup_in_contract(contract_id)
-            {
-                return Some(info);
-            }
-        }
-
-        // Case 2: base is a type conversion call like ICallee(addr)
-        if let ExprKind::Call(callee, _args, _named) = &base.kind
-            && let ExprKind::Ident(res_slice) = &callee.kind
-            && let Some(hir::Res::Item(hir::ItemId::Contract(contract_id))) = res_slice.first()
-            && let Some(info) = lookup_in_contract(*contract_id)
-        {
-            return Some(info);
-        }
-
-        // Case 3: base is `this` (Builtin::This)
-        if let ExprKind::Ident(res_slice) = &base.kind
-            && let Some(hir::Res::Builtin(Builtin::This)) = res_slice.first()
-        {
-            for contract_id in self.gcx.hir.contract_ids() {
-                if let Some(info) = lookup_in_contract(contract_id) {
-                    return Some(info);
-                }
-            }
-        }
-
-        None
-    }
-
-    /// Resolves a member of a builtin module to its corresponding builtin.
-    /// For example, (Builtin::Msg, "sender") -> Some(Builtin::MsgSender)
-    fn resolve_builtin_member(&self, base: Builtin, member: Symbol) -> Option<Builtin> {
-        match base {
-            Builtin::Msg => {
-                if member == sym::sender {
-                    Some(Builtin::MsgSender)
-                } else if member == sym::value {
-                    Some(Builtin::MsgValue)
-                } else if member == sym::data {
-                    Some(Builtin::MsgData)
-                } else if member == sym::sig {
-                    Some(Builtin::MsgSig)
-                } else if member == kw::Gas {
-                    Some(Builtin::MsgGas)
-                } else {
-                    None
-                }
-            }
-            Builtin::Block => {
-                if member == kw::Coinbase {
-                    Some(Builtin::BlockCoinbase)
-                } else if member == kw::Timestamp {
-                    Some(Builtin::BlockTimestamp)
-                } else if member == kw::Difficulty {
-                    Some(Builtin::BlockDifficulty)
-                } else if member == kw::Prevrandao {
-                    Some(Builtin::BlockPrevrandao)
-                } else if member == kw::Number {
-                    Some(Builtin::BlockNumber)
-                } else if member == kw::Gaslimit {
-                    Some(Builtin::BlockGaslimit)
-                } else if member == kw::Chainid {
-                    Some(Builtin::BlockChainid)
-                } else if member == kw::Basefee {
-                    Some(Builtin::BlockBasefee)
-                } else if member == kw::Blobbasefee {
-                    Some(Builtin::BlockBlobbasefee)
-                } else {
-                    None
-                }
-            }
-            Builtin::Tx => {
-                if member == kw::Origin {
-                    Some(Builtin::TxOrigin)
-                } else if member == kw::Gasprice {
-                    Some(Builtin::TxGasPrice)
-                } else {
-                    None
-                }
-            }
-            Builtin::Abi => {
-                if member == sym::encode {
-                    Some(Builtin::AbiEncode)
-                } else if member == sym::encodePacked {
-                    Some(Builtin::AbiEncodePacked)
-                } else if member == sym::encodeWithSelector {
-                    Some(Builtin::AbiEncodeWithSelector)
-                } else if member == sym::encodeCall {
-                    Some(Builtin::AbiEncodeCall)
-                } else if member == sym::encodeWithSignature {
-                    Some(Builtin::AbiEncodeWithSignature)
-                } else if member == sym::decode {
-                    Some(Builtin::AbiDecode)
-                } else {
-                    None
-                }
-            }
-            _ => None,
-        }
-    }
-
-    /// Finds a library function by name, preferring the overload matching arg_count.
-    pub(super) fn find_library_function(
-        &self,
-        library_id: hir::ContractId,
-        name: Symbol,
-        arg_count: usize,
-    ) -> Option<hir::FunctionId> {
-        let library = self.gcx.hir.contract(library_id);
-        let mut candidates: Vec<hir::FunctionId> = Vec::new();
-
-        for func_id in library.all_functions() {
-            let func = self.gcx.hir.function(func_id);
-            if func.name.is_some_and(|n| n.name == name) {
-                candidates.push(func_id);
-            }
-        }
-
-        // Select the overload that matches the argument count
-        candidates
-            .iter()
-            .find(|&&fid| {
-                let f = self.gcx.hir.function(fid);
-                f.parameters.len() == arg_count
-            })
-            .copied()
-            .or_else(|| candidates.first().copied())
-    }
-
-    /// Lowers an internal function call by inlining it.
-    /// This handles calls like `add(a, b)` where `add` is a function in the same contract.
-    fn lower_internal_call(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        func_id: hir::FunctionId,
-        args: &CallArgs<'_>,
-    ) -> ValueId {
-        let func = self.gcx.hir.function(func_id);
-
-        // Collect argument values FIRST (before entering inline tracking)
-        // This allows nested calls to the same function (e.g., add(add(x, 1), 2))
-        // because we evaluate arguments before marking ourselves as "in progress"
-        let arg_vals: Vec<ValueId> =
-            args.exprs().map(|arg| self.lower_expr(builder, arg)).collect();
-
-        if func.returns.is_empty() {
-            if self.function_is_recursive(func_id) {
-                return self.lower_internal_call_fallback(builder, func_id, arg_vals);
-            }
-            return self.lower_inline_void_call(builder, func_id, arg_vals);
-        }
-
-        // The SSA inline path (`lower_library_body_simple`) only models a
-        // straight-line body that ends in a `return`. Anything else — a loop, an
-        // `if`, a multi-statement control flow — is lowered as a real
-        // `internal_call` instead, where the memory-backed internal frame handles
-        // reassigned locals, loops, and recursion correctly. Recursive functions
-        // with a simple ternary body (which `is_simple_return_function` accepts)
-        // are caught separately so inlining does not hit a recursive cycle.
-        // Simple, non-recursive functions still inline. Internal/private callees
-        // use the internal-frame convention directly; a public callee is compiled
-        // for the external ABI, so it needs an internal-frame copy
-        // (`ensure_internal_mir_function`) for `internal_call` to target.
-        let needs_call =
-            !Self::is_simple_return_function(func) || self.function_is_recursive(func_id);
-        if needs_call {
-            return self.lower_internal_call_fallback(builder, func_id, arg_vals);
-        }
-
-        // Check for recursive inlining cycle AFTER evaluating arguments.
-        if !self.try_enter_inline(func_id) {
-            return self.lower_internal_call_fallback(builder, func_id, arg_vals);
-        }
-
-        // Save current locals
-        let saved_locals = std::mem::take(&mut self.locals);
-
-        // Bind parameters to argument values directly (SSA style)
-        for (i, &param_id) in func.parameters.iter().enumerate() {
-            if let Some(&arg_val) = arg_vals.get(i) {
-                self.locals.insert(param_id, arg_val);
-            }
-        }
-
-        // For simple functions with a single return statement, extract and evaluate directly
-        let result = if let Some(body) = &func.body {
-            self.lower_library_body_simple(builder, body, func)
-        } else {
-            builder.imm_u64(0)
-        };
-
-        // Restore locals
-        self.locals = saved_locals;
-
-        // Exit inline tracking
-        self.exit_inline();
-
-        result
-    }
-
-    fn lower_internal_call_fallback(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        func_id: hir::FunctionId,
-        arg_vals: Vec<ValueId>,
-    ) -> ValueId {
-        let func = self.gcx.hir.function(func_id);
-        let result_ty = func
-            .returns
-            .first()
-            .map(|&ret_id| self.lower_type_from_var(self.gcx.hir.variable(ret_id)));
-        let is_internal =
-            matches!(func.visibility, hir::Visibility::Internal | hir::Visibility::Private);
-        let mir_id = if is_internal {
-            self.ensure_function_lowered(func_id)
-        } else {
-            self.ensure_internal_mir_function(func_id)
-        };
-        builder.internal_call(mir_id, arg_vals, result_ty, func.returns.len())
-    }
-
-    /// Lowers a base constructor call using already-resolved constructor arguments.
-    pub(super) fn lower_base_constructor_call(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        ctor_id: hir::FunctionId,
-        modifier: Option<&hir::Modifier<'_>>,
-    ) -> ValueId {
-        let ctor = self.gcx.hir.function(ctor_id);
-        let arg_exprs: Vec<_> = modifier.map(|m| m.args.exprs().collect()).unwrap_or_default();
-        let arg_vals: Vec<ValueId> = ctor
-            .parameters
-            .iter()
-            .enumerate()
-            .map(|(i, &param_id)| {
-                let param = self.gcx.hir.variable(param_id);
-                if let Some(arg) = arg_exprs.get(i) {
-                    self.lower_constructor_arg(builder, arg, &param.ty)
-                } else {
-                    builder.imm_u64(0)
-                }
-            })
-            .collect();
-
-        self.lower_inline_void_call(builder, ctor_id, arg_vals)
-    }
-
-    /// Lowers a void internal function by inlining its full statement body.
-    fn lower_inline_void_call(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        func_id: hir::FunctionId,
-        arg_vals: Vec<ValueId>,
-    ) -> ValueId {
-        let func = self.gcx.hir.function(func_id);
-        let parameters: Vec<_> = func.parameters.to_vec();
-        let body = func.body;
-
-        if !self.try_enter_inline(func_id) {
-            panic!("codegen hit unsupported void-call inline recursion");
-        }
-
-        let saved_locals = std::mem::take(&mut self.locals);
-        let saved_local_memory_slots = std::mem::take(&mut self.local_memory_slots);
-        let saved_next_local_memory_offset = self.next_local_memory_offset;
-        let saved_assigned_vars = std::mem::take(&mut self.assigned_vars);
-
-        if let Some(body) = body {
-            self.collect_assigned_vars_block(&body);
-        }
-
-        for (i, param_id) in parameters.into_iter().enumerate() {
-            if let Some(&arg_val) = arg_vals.get(i) {
-                self.locals.insert(param_id, arg_val);
-            }
-        }
-
-        if let Some(body) = body {
-            let saved_in_unchecked_block = self.in_unchecked_block;
-            self.in_unchecked_block = false;
-            self.lower_block(builder, &body);
-            self.in_unchecked_block = saved_in_unchecked_block;
-        }
-
-        self.locals = saved_locals;
-        self.local_memory_slots = saved_local_memory_slots;
-        self.next_local_memory_offset = saved_next_local_memory_offset;
-        self.assigned_vars = saved_assigned_vars;
-        self.exit_inline();
-
-        builder.imm_u64(0)
-    }
-
-    /// Lowers constructor arguments into the representation expected by the
-    /// callee body. Memory `bytes`/`string` parameters receive Solidity's
-    /// `[length][data...]` memory pointer, including literal base-constructor
-    /// arguments such as `ERC20("Name", "SYM")`.
-    fn lower_constructor_arg(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        arg: &hir::Expr<'_>,
-        param_ty: &hir::Type<'_>,
-    ) -> ValueId {
-        if matches!(
-            param_ty.kind,
-            hir::TypeKind::Elementary(hir::ElementaryType::String | hir::ElementaryType::Bytes)
-        ) {
-            return self.lower_expr_as_memory_bytes(builder, arg);
-        }
-
-        self.lower_expr(builder, arg)
-    }
-
-    /// Lowers an internal library function call by inlining it.
-    /// For internal library functions, we inline the function body.
-    fn lower_library_call(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        func_id: hir::FunctionId,
-        args: &CallArgs<'_>,
-        bound_arg: Option<ValueId>,
-    ) -> ValueId {
-        let func = self.gcx.hir.function(func_id);
-
-        // For internal library functions, inline the function body
-        if func.visibility == hir::Visibility::Internal
-            || func.visibility == hir::Visibility::Private
-        {
-            // Collect argument values FIRST (before entering inline tracking)
-            // This allows nested calls to the same function (e.g., add(add(x, 1), 2))
-            // because we evaluate arguments before marking ourselves as "in progress"
-            let mut arg_vals: Vec<ValueId> = Vec::new();
-
-            // If there's a bound argument (from `using X for T`), it's the first argument
-            if let Some(bound_val) = bound_arg {
-                arg_vals.push(bound_val);
-            }
-
-            // Lower all explicit arguments
-            for arg in args.exprs() {
-                arg_vals.push(self.lower_expr(builder, arg));
-            }
-
-            if func.returns.is_empty() {
-                if self.function_is_recursive(func_id) {
-                    return self.lower_internal_call_fallback(builder, func_id, arg_vals);
-                }
-                return self.lower_inline_void_call(builder, func_id, arg_vals);
-            }
-
-            if !Self::is_simple_return_function(func) || self.function_is_recursive(func_id) {
-                return self.lower_internal_call_fallback(builder, func_id, arg_vals);
-            }
-
-            // Check for recursive inlining cycle AFTER evaluating arguments.
-            if !self.try_enter_inline(func_id) {
-                return self.lower_internal_call_fallback(builder, func_id, arg_vals);
-            }
-
-            // Simple inlining: bind parameters directly as SSA values
-            // This works for pure functions that don't mutate parameters
-            // Save current locals
-            let saved_locals = std::mem::take(&mut self.locals);
-            let saved_local_memory_slots = std::mem::take(&mut self.local_memory_slots);
-            let saved_next_local_memory_offset = self.next_local_memory_offset;
-            let saved_assigned_vars = std::mem::take(&mut self.assigned_vars);
-
-            if let Some(body) = &func.body {
-                self.collect_assigned_vars_block(body);
-            }
-
-            // Bind parameters to argument values directly (SSA style)
-            for (i, &param_id) in func.parameters.iter().enumerate() {
-                if let Some(&arg_val) = arg_vals.get(i) {
-                    self.locals.insert(param_id, arg_val);
-                }
-            }
-
-            // For simple functions with a single return statement, extract and evaluate directly
-            let result = if let Some(body) = &func.body {
-                self.lower_library_body_simple(builder, body, func)
-            } else {
-                builder.imm_u64(0)
-            };
-
-            // Restore locals
-            self.locals = saved_locals;
-            self.local_memory_slots = saved_local_memory_slots;
-            self.next_local_memory_offset = saved_next_local_memory_offset;
-            self.assigned_vars = saved_assigned_vars;
-
-            // Exit inline tracking
-            self.exit_inline();
-
-            result
-        } else {
-            panic!("codegen does not support external library calls yet")
-        }
-    }
-
-    fn is_simple_return_function(func: &hir::Function<'_>) -> bool {
-        if func.returns.len() != 1 {
-            return false;
-        }
-        let Some(body) = func.body else {
-            return false;
-        };
-        body.stmts.iter().any(|stmt| matches!(stmt.kind, hir::StmtKind::Return(Some(_))))
-            && body.stmts.iter().all(|stmt| {
-                matches!(
-                    stmt.kind,
-                    hir::StmtKind::DeclSingle(_)
-                        | hir::StmtKind::Expr(_)
-                        | hir::StmtKind::Return(Some(_))
-                )
-            })
-    }
-
-    /// Whether `func_id` directly or indirectly calls itself (cached). A recursive function
-    /// must be lowered as a real `internal_call` instead of being inlined.
-    fn function_is_recursive(&mut self, func_id: hir::FunctionId) -> bool {
-        if let Some(&cached) = self.recursive_functions.get(&func_id) {
-            return cached;
-        }
-        let mut visiting = FxHashSet::default();
-        let result = self.function_reaches(func_id, func_id, &mut visiting);
-        self.recursive_functions.insert(func_id, result);
-        result
-    }
-
-    fn function_reaches(
-        &self,
-        current: hir::FunctionId,
-        target: hir::FunctionId,
-        visiting: &mut FxHashSet<hir::FunctionId>,
-    ) -> bool {
-        if !visiting.insert(current) {
-            return false;
-        }
-
-        for callee in self.function_callees(current) {
-            if callee == target || self.function_reaches(callee, target, visiting) {
-                return true;
-            }
-        }
-
-        false
-    }
-
-    fn function_callees(&self, func_id: hir::FunctionId) -> Vec<hir::FunctionId> {
-        let mut callees = Vec::new();
-        let func = self.gcx.hir.function(func_id);
-        if let Some(body) = func.body {
-            for stmt in body.stmts {
-                self.stmt_collect_callees(stmt, &mut callees);
-            }
-        }
-        callees
-    }
-
-    /// Collects calls contained recursively in a statement.
-    fn stmt_collect_callees(&self, stmt: &hir::Stmt<'_>, callees: &mut Vec<hir::FunctionId>) {
-        use hir::StmtKind;
-        match &stmt.kind {
-            StmtKind::Expr(e)
-            | StmtKind::Return(Some(e))
-            | StmtKind::Revert(e)
-            | StmtKind::Emit(e) => self.expr_collect_callees(e, callees),
-            StmtKind::Block(b) | StmtKind::UncheckedBlock(b) | StmtKind::AssemblyBlock(b) => {
-                for stmt in b.stmts {
-                    self.stmt_collect_callees(stmt, callees);
-                }
-            }
-            StmtKind::If(c, t, e) => {
-                self.expr_collect_callees(c, callees);
-                self.stmt_collect_callees(t, callees);
-                if let Some(e) = e {
-                    self.stmt_collect_callees(e, callees);
-                }
-            }
-            StmtKind::Loop(b, _) => {
-                for stmt in b.stmts {
-                    self.stmt_collect_callees(stmt, callees);
-                }
-            }
-            StmtKind::Switch(sw) => {
-                self.expr_collect_callees(sw.selector, callees);
-                for case in sw.cases {
-                    for stmt in case.body.stmts {
-                        self.stmt_collect_callees(stmt, callees);
-                    }
-                }
-            }
-            StmtKind::Try(t) => {
-                self.expr_collect_callees(&t.expr, callees);
-                for clause in t.clauses {
-                    for stmt in clause.block.stmts {
-                        self.stmt_collect_callees(stmt, callees);
-                    }
-                }
-            }
-            StmtKind::DeclSingle(var_id) => {
-                if let Some(init) = self.gcx.hir.variable(*var_id).initializer {
-                    self.expr_collect_callees(init, callees);
-                }
-            }
-            StmtKind::DeclMulti(_, init) => self.expr_collect_callees(init, callees),
-            StmtKind::Return(None)
-            | StmtKind::Continue
-            | StmtKind::Break
-            | StmtKind::Placeholder
-            | StmtKind::Err(_) => {}
-        }
-    }
-
-    /// Collects calls contained recursively in an expression.
-    fn expr_collect_callees(&self, expr: &hir::Expr<'_>, callees: &mut Vec<hir::FunctionId>) {
-        match &expr.kind {
-            ExprKind::Call(callee, args, _) => {
-                if let ExprKind::Ident(res) = &callee.kind
-                    && let Some(hir::Res::Item(hir::ItemId::Function(f))) = res.first()
-                {
-                    callees.push(*f);
-                }
-                self.expr_collect_callees(callee, callees);
-                for arg in args.exprs() {
-                    self.expr_collect_callees(arg, callees);
-                }
-            }
-            ExprKind::Binary(l, _, r) | ExprKind::Assign(l, _, r) => {
-                self.expr_collect_callees(l, callees);
-                self.expr_collect_callees(r, callees);
-            }
-            ExprKind::Unary(_, e)
-            | ExprKind::Member(e, _)
-            | ExprKind::YulMember(e, _)
-            | ExprKind::Payable(e)
-            | ExprKind::Delete(e) => self.expr_collect_callees(e, callees),
-            ExprKind::Ternary(c, t, f) => {
-                self.expr_collect_callees(c, callees);
-                self.expr_collect_callees(t, callees);
-                self.expr_collect_callees(f, callees);
-            }
-            ExprKind::Index(b, i) => {
-                self.expr_collect_callees(b, callees);
-                if let Some(i) = i {
-                    self.expr_collect_callees(i, callees);
-                }
-            }
-            ExprKind::Slice(b, s, e) => {
-                self.expr_collect_callees(b, callees);
-                if let Some(s) = s {
-                    self.expr_collect_callees(s, callees);
-                }
-                if let Some(e) = e {
-                    self.expr_collect_callees(e, callees);
-                }
-            }
-            ExprKind::Array(es) => {
-                for e in *es {
-                    self.expr_collect_callees(e, callees);
-                }
-            }
-            ExprKind::Tuple(es) => {
-                for e in es.iter().flatten() {
-                    self.expr_collect_callees(e, callees);
-                }
-            }
-            ExprKind::New(_)
-            | ExprKind::TypeCall(_)
-            | ExprKind::Lit(_)
-            | ExprKind::Ident(_)
-            | ExprKind::Type(_)
-            | ExprKind::Err(_) => {}
-        }
-    }
-
-    /// Lowers a simple library function body.
-    /// For functions with a single return statement, directly evaluate the return expression.
-    fn lower_library_body_simple(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        body: &hir::Block<'_>,
-        func: &hir::Function<'_>,
-    ) -> ValueId {
-        let saved_in_unchecked_block = self.in_unchecked_block;
-        self.in_unchecked_block = false;
-
-        for &return_id in func.returns {
-            let zero = builder.imm_u64(0);
-            self.locals.insert(return_id, zero);
-        }
-
-        let result = if let Some(value) = self.lower_library_block_return(builder, body) {
-            value
-        } else {
-            // Implicit named returns: the body assigned the named return variables
-            // (e.g. `success = ...; result = ...;` with no explicit `return`). Write
-            // returns 1..N to scratch memory at offset `i * 32` so the caller's
-            // `lower_multi_var_decl` (which reads `mload(i * 32)`) recovers them; the
-            // first return flows back as the MIR value below.
-            if func.returns.len() > 1 {
-                for (i, &return_id) in func.returns.iter().enumerate().skip(1) {
-                    if let Some(&value) = self.locals.get(&return_id) {
-                        let offset = builder.imm_u64((i * 32) as u64);
-                        builder.mstore(offset, value);
-                    }
-                }
-            }
-
-            if let Some(&return_id) = func.returns.first()
-                && let Some(&value) = self.locals.get(&return_id)
-            {
-                value
-            } else {
-                builder.imm_u64(0)
-            }
-        };
-
-        self.in_unchecked_block = saved_in_unchecked_block;
-        result
-    }
-
-    fn lower_library_block_return(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        block: &hir::Block<'_>,
-    ) -> Option<ValueId> {
-        for stmt in block.stmts {
-            if let Some(value) = self.lower_library_stmt_return(builder, stmt) {
-                return Some(value);
-            }
-        }
-        None
-    }
-
-    /// Extract return value from a statement after lowering prior side effects in that statement.
-    fn lower_library_stmt_return(
-        &mut self,
-        builder: &mut FunctionBuilder<'_>,
-        stmt: &hir::Stmt<'_>,
-    ) -> Option<ValueId> {
-        match &stmt.kind {
-            hir::StmtKind::Return(Some(expr)) => Some(self.lower_expr(builder, expr)),
-            hir::StmtKind::Return(None) => Some(builder.imm_u64(0)),
-            hir::StmtKind::DeclSingle(var_id) => {
-                let var = self.gcx.hir.variable(*var_id);
-                let init_val = if let Some(init) = var.initializer {
-                    self.lower_expr(builder, init)
-                } else {
-                    builder.imm_u64(0)
-                };
-                self.locals.insert(*var_id, init_val);
-                None
-            }
-            hir::StmtKind::Expr(expr) => {
-                self.lower_expr(builder, expr);
-                None
-            }
-            hir::StmtKind::Block(block) => self.lower_library_block_return(builder, block),
-            hir::StmtKind::UncheckedBlock(block) => self.lower_library_block_return(builder, block),
-            hir::StmtKind::If(cond, then_stmt, else_stmt) => {
-                let cond_val = self.lower_expr(builder, cond);
-                let then_return = self.lower_library_stmt_return(builder, then_stmt);
-                let else_return =
-                    else_stmt.map(|else_stmt| self.lower_library_stmt_return(builder, else_stmt));
-
-                match (then_return, else_return.flatten()) {
-                    (Some(then_val), Some(else_val)) => {
-                        Some(builder.select(cond_val, then_val, else_val))
-                    }
-                    // A one-sided return is an early-return control-flow shape. This helper
-                    // returns expression values only, so let later statements provide the
-                    // fallthrough value instead of treating the branch as unconditional.
-                    _ => None,
-                }
-            }
-            hir::StmtKind::DeclMulti(vars, rhs) => {
-                self.lower_multi_var_decl(builder, vars, rhs);
-                None
-            }
-            hir::StmtKind::Loop(..)
-            | hir::StmtKind::AssemblyBlock(_)
-            | hir::StmtKind::Switch(_)
-            | hir::StmtKind::Emit(_)
-            | hir::StmtKind::Revert(_)
-            | hir::StmtKind::Break
-            | hir::StmtKind::Continue
-            | hir::StmtKind::Try(_)
-            | hir::StmtKind::Placeholder
-            | hir::StmtKind::Err(_) => {
-                self.lower_stmt(builder, stmt);
-                None
-            }
-        }
-    }
-
-    /// Checks if an expression has a contract type (as opposed to address type).
-    /// Used to distinguish between address.transfer(amount) and token.transfer(to, amount).
-    pub(super) fn is_contract_type_expr(&self, expr: &hir::Expr<'_>) -> bool {
-        // Case 1: Variable with contract type (e.g., `token` where `MinimalERC20 token`)
-        if let ExprKind::Ident(res_slice) = &expr.kind
-            && let Some(hir::Res::Item(hir::ItemId::Variable(var_id))) = res_slice.first()
-        {
-            let var = self.gcx.hir.variable(*var_id);
-            let ty = self.gcx.type_of_hir_ty(&var.ty);
-            if matches!(ty.kind, solar_sema::ty::TyKind::Contract(_)) {
-                return true;
-            }
-        }
-
-        // Case 2: Type conversion call like IToken(addr)
-        if let ExprKind::Call(callee, _, _) = &expr.kind
-            && let ExprKind::Ident(res_slice) = &callee.kind
-            && let Some(hir::Res::Item(hir::ItemId::Contract(_))) = res_slice.first()
-        {
-            return true;
-        }
-
-        false
-    }
-
-    /// Resolves a member call that may be a `using X for Y` library extension call.
-    ///
-    /// If `base.member` matches a using directive, returns the library function ID.
-    /// The caller should pass the evaluated `base` as the first (bound) argument.
-    pub(super) fn resolve_using_directive_call(
-        &self,
-        base: &hir::Expr<'_>,
-        member_name: Symbol,
-    ) -> Option<hir::FunctionId> {
-        let contract_id = self.current_contract_id?;
-        let contract = self.gcx.hir.contract(contract_id);
-
-        // Get the type of the base expression
-        let base_ty = self.get_expr_type(base)?;
-
-        // Search through source- and contract-level using directives.
-        for using in self.gcx.hir.source(contract.source).usings.iter().chain(contract.usings) {
-            // Check if this using directive applies to the base type
-            let type_matches = if let Some(ref target_ty) = using.ty {
-                // Check if the types match
-                self.types_match_for_using(&base_ty, target_ty)
-            } else {
-                // `using X for *` matches all types
-                true
-            };
-
-            if !type_matches {
-                continue;
-            }
-
-            for entry in using.entries {
-                if entry.operator.is_some() {
-                    continue;
-                }
-
-                match &entry.kind {
-                    hir::UsingEntryKind::Library(library) => {
-                        for func_id in self.gcx.hir.contract(*library).functions() {
-                            let func = self.gcx.hir.function(func_id);
-                            if func.name.map(|n| n.name) != Some(member_name) {
-                                continue;
-                            }
-                            if self.using_function_matches(&base_ty, func_id) {
-                                return Some(func_id);
-                            }
-                        }
-                    }
-                    hir::UsingEntryKind::Functions(functions) => {
-                        for &func_id in *functions {
-                            let func = self.gcx.hir.function(func_id);
-                            let name = entry.name.or_else(|| func.name.map(|n| n.name));
-                            if name != Some(member_name) {
-                                continue;
-                            }
-                            if self.using_function_matches(&base_ty, func_id) {
-                                return Some(func_id);
-                            }
-                        }
-                    }
-                    hir::UsingEntryKind::Err(_) => {}
-                }
-            }
-        }
-
-        None
-    }
-
-    pub(super) fn using_function_matches(
-        &self,
-        base_ty: &solar_sema::ty::Ty<'_>,
-        func_id: hir::FunctionId,
-    ) -> bool {
-        let func = self.gcx.hir.function(func_id);
-        if !matches!(func.visibility, hir::Visibility::Internal | hir::Visibility::Private) {
-            return false;
-        }
-
-        let Some(&first_param_id) = func.parameters.first() else {
-            return false;
-        };
-        let first_param = self.gcx.hir.variable(first_param_id);
-        self.types_match_for_using(base_ty, &first_param.ty)
     }
 }
