+++ crates/codegen/src/lower/mod.rs
@@ -5,6 +5,7 @@
 mod abi_encode;
 mod abi_packed;
 mod bytes;
+mod call;
 mod checked_arith;
 mod expr;
 mod index;
